Modules/MessagePassing/requirements.txt
functions-framework==3.*
google-cloud-firestore
google-cloud-pubsub
boto3
protobuf
requests